backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase==2.18.1
psycopg2-binary==2.9.9

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Processing
pandas==2.3.2
openpyxl==3.1.5

# Email & Scheduling
yagmail==0.15.293
schedule==1.2.0
pytz==2023.3

# Development & Testing
pytest==7.4.3
python-dotenv==1.1.1
websockets>=11.0

# CORS
fastapi-cors==0.0.6

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
